requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
google-cloud-bigquery==3.17.0
google-auth==2.27.0
plotly==5.18.0
